Relocating the eventloop to a standalone file
