+++ simulator/node/mobility.py
@@ -1,7 +1,7 @@
 import math
 from typing import Dict
 
-from simulator.event.loop import EventLoop
+from simulator.event import EventLoop
 from simulator.messages.mobility import MobilityCommand, MobilityCommandType
 from simulator.node import Node, Position
 from simulator.node.interface import INodeHandler
